% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{R Shiny app to create simplified TS domain dataset}
\arguments{
\item{fn}{a file name or URL pointing to script metadata file}
}
\value{
R shiny code for providing inputs and downloading TS file
}
\description{
Create simplified Trial Summary (TS) domain dataset
}
\examples{
\dontrun{
  install.package(genTS)
  b <- "development/R/scripts"
  c <- "Draw_Dist2_R.yml"
  f1 <- paste(a,b,c, sep = '/')
  r1 <- build_inputs(f1)
}
}
\author{
Hanming Tu
}
